+++ app/build.gradle.kts
@@ -108,4 +108,5 @@ dependencies {
 
     // Compose Navigation
     implementation ("androidx.navigation:navigation-compose:2.4.0-alpha10")
+
 }
